+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+<a name="v0.0.5"></a>
+### v0.0.5 (2013-09-30)
+
+
+#### Bug Fixes
+
+* **desktop:** unbind keydown when scroll elm is destroyed ([51500ce9](http://github.com/ajoslin/angular-scrolly/commit/51500ce93087f5e62e84207e465722710afc3d17))
+* **dragger:** stop propagation, so scroll on popups works ([2006b0bc](http://github.com/ajoslin/angular-scrolly/commit/2006b0bc658c78d4349ef718b860c557cd3a7e97))
+* **scroller:**
+  * remove scroll event ([c8b170e2](http://github.com/ajoslin/angular-scrolly/commit/c8b170e29728d488ec4a312c19a9e7f60f07245b))
+  * count container top and bottom offset ([f8b71798](http://github.com/ajoslin/angular-scrolly/commit/f8b71798c7acfddf8dca6912f50245479ecdc358))
+* **transformer:** Try both cases of transformProp (firefox, ie10) ([761f7c22](http://github.com/ajoslin/angular-scrolly/commit/761f7c22450c6c56cd96c079ec658ddcfcef3044))
+
+
+#### Features
+
+* **$scroller:** allow desktop scrolling ([8249eaf2](http://github.com/ajoslin/angular-scrolly/commit/8249eaf25049025a014e9439e8febb0b5e190100))
+* **dragger:**
+  * Give 2d x and y data instead of 1d ([559ea186](http://github.com/ajoslin/angular-scrolly/commit/559ea1861996a594e7bdbc0642e147fc29640ff0))
+  * add horizontal option ([ddd75112](http://github.com/ajoslin/angular-scrolly/commit/ddd75112c88b3616f9d313d8ff233dcdbb2763f2))
+* **scroller:** support mousewheel, keys ([a365debd](http://github.com/ajoslin/angular-scrolly/commit/a365debd03e7fafcc2fe83bc34d0fbdaebbe4074))
+
 <a name="v0.0.4"></a>
 ### v0.0.4 (2013-08-13)